+++ source/win-preferences/schema/editor.ts
@@ -124,6 +124,12 @@ export function getEditorFields (config: ConfigOptions): PreferencesFieldset[] {
               label: trans('Render pandoc divs and spans'),
               model: 'display.renderPandoc',
               disabled: config.display.renderingMode === 'raw'
+            },
+            {
+              type: 'checkbox',
+              label: trans('Render horizontal rules'),
+              model: 'display.renderHorizontalRules',
+              disabled: config.display.renderingMode === 'raw'
             }
           ]
         }